Check IPv6 status as part of firewall check
Case CPANEL-37936: In order to verify that the MySQL port was blocked in
the firewall, we would check that it was blocked for both IPv4 and IPv6
addresses.  However, this could result in us giving bad advice in the
event that IPv6 was disabled in the kernel since there would be no need
to block the MySQL port for IPv6 in that instance.  This change ensures
that IPv6 is enabled as part of the firewall check.

Changelog: Check IPv6 status as part of firewall check