Adjust notice about removal of ClamAV when ImunifyAV is installed.
Case CPANEL-34196: When ImunifyAV (free version) was added to
the product, the Security Advisor was updated to check whether
ClamAV was also installed, and if so, recommend removing it on
the grounds that it would be redundant. ClamAV is not redundant,
however, because it is still required for a setting in WHM that
enables scanning messages that pass through Exim. Instead of plainly
recommending ClamAV's removal, we should add some detail on why it
might (or might not) still be necessary.

This change also updates the RPM check for compatibility with versions
98 and up, which no longer have the Cpanel::RPM module, and fixes
several broken unit tests.