Update link to MultiPHP Manager interface
Case CPANEL-38967: The MultiPHP Manager interface was converted to an
angular application and had its URI updated as part of the conversion.
However, Security Advisor was not made aware of the new URI and was
still providing the link to the deprecated/removed interface which
resulted in a 404.  This change updates the link to point to the new
URI.

Changelog: Update link to MultiPHP Manager interface